requirements.txt
python-telegram-bot[webhooks]==21.9
gspread==5.12.0
google-auth==2.23.4
python-dotenv==1.0.0